refactor landing page and add routing
